Better layout division
