+++ Site/src/lib/components/DelfaultFooter.svelte
@@ -1,0 +1,14 @@
+<footer class="flex flex-col items-center gap-1 py-6 border-t border-slate-800">
+    <div class="flex gap-3">
+        <span>&copy; 2023 - Mateus Ryan</span>
+        <span class="text-slate-800">&#x2022;</span>
+        <a class="text-slate-300 visited:text-slate-300 hover:text-sky-700" href="https://github.com/Mth-Ryan" target="_blank">
+            Github
+        </a>
+        <span class="text-slate-800">&#x2022;</span>
+        <a class="text-slate-300 visited:text-slate-300 hover:text-sky-700" href="https://linkedin.com/in/mateus-ryan-aa63b3210" target="_blank">
+            Linkedin
+        </a>
+    </div>
+    <span>Licenced under <a class="visited:text-sky-700" href="https://www.gnu.org/licenses/agpl-3.0.html">AGPL-3.0</a></span>
+</footer>
